requirements.txt
dedalus-labs==0.1.1
fastapi==0.115.2
huggingface-hub==1.1.4
httpx==0.28.1
python-dotenv==1.2.1
pytest==8.3.3
requests==2.32.5
uvicorn==0.32.0
